pyproject.toml
[project]
name = "moonshot-mcp-server"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "colorlog>=6.9.0",
    "mcp[cli]>=1.9.0",
    "pydantic>=2.11.3",
    "python-dotenv>=1.0.1",
    "pydantic-settings>=2.8.1",
    "ruff>=0.11.2",
    "pre-commit>=4.2.0",
    "toml",
    "tomli>=2.2.1",
    "watchdog>=6.0.0",
    "uvicorn>=0.34.1",
    "fastmcp>=2.3.4",
    "fastapi[standard]>=0.115.12",
    "polib>=1.2.0",
]
